convert to int

+++ models/customer_dim_customers.sql
@@ -1,2 +1,2 @@
 select * from {{ ref('tpch', 'dim_customers') }}
-where customer_key = split(current_user, '_')[1]
+where customer_key = split(current_user, '_')[1]::int
