+++ models/customer_fct_order_items.sql
@@ -1,2 +1,2 @@
 select * from {{ ref('tpch', 'fct_order_items') }}
-where customer_key = split(current_user, '_')[1]
+where customer_key = split(current_user, '_')[1]::int
